Move GithubApi into separate file
